kvstore and sqlite benchmarks runnable; iterator implemented; sequential read implemented;

+++ common/persist.hh
@@ -28,7 +28,7 @@ inline void mfence() {
   asm volatile("mfence":::"memory");
 }
 
-inline void clflush(char *data, int len) {
+inline void clflush(char *data, size_t len) {
   if (data == NULL) return;
   volatile char *ptr = (char *)((unsigned long)data &~(CACHE_LINE_SIZE-1));
   mfence();
